Fix representation Object Oid
